fix: CMP-10579 Remove error iframe
React adds an iframe over the top of the component whenver there is an
error. This is annoying because it prevents the extension from working and
can show for non-issues e.g. missing Gainsight ID. The style has been set
to pointer-events: none so it doesn't block use of the extension